Handle disconnections we haven't caused
When the XWiiIface Dispatch method returns an event of XWII_EVENT_WATCH,
this means that the board has disconnected (as we haven't enabled
hotplug events).  Use this information to send the DISCONNECTED status
to the UI if the board disconnects for any reason, not just when we have
told it to.

Also tidy up the connect / disconnect code to be driven by received
messages instead of using a flag in the board event loop.