Stop wii-scale using a CPU core when connected
Adjust the XWiiIface class so that ::Dispatch only returns when there is
an event to deal with and make it use poll() internally so that it
doesn't use any CPU unless the balance board has sent an event.  This
stops the wii-scale process using 100% of a CPU core in-between received
events by continually calling xwii_iface_dispatch().