Make skipping values at the start of weigh dynamic
Instead of discarding a fixed number of values at the start of weighing
(10), calculate the standard deviation of all recorded values and
discard values smaller than the mean minus two standard deviations at
the start.  This automatically compensates for fast or slow stepping on
to the balance board and should make for more consistent results.