Make weighing process twice as fast
The back end waits for 100 milliseconds between sending readings so
that it doesn't overwhelm the front-end with readings and complete the
weighing process too quickly.  However, 100ms * 50 readings = 5 seconds
which is a fair bit slower than the 3 seconds mentioned in the README.

To align the weighing speed with the user's expectations from the
README, halve the wait between sending readings to 50ms so that the
overall process takes 2.5s.