Work around Bluetooth address missing from sysfs
On some machines (possibly kernel specific), the 'address' attribute
containing the MAC address of the balance board is missing from the
parent of the balance board hid device under /sys/devices/, which causes
the backend to throw an exception on connect.

To work around this, fall back to finding the first connected balance
board and disconnect this (as it is very unlikely that someone will have
two balance boards connected at once).